refactor: remove cli/prepare module
